development/requirements.txt
Django==2.2.13
djangorestframework==3.9.1
mysqlclient==2.0.1
openapi-codec==1.3.2
requests==2.21.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
uWSGI==2.0.17.1
unicodecsv==0.14.1
pep8==1.7.1
pylint==2.5.2
pandas==1.0.3
celery==4.3.0
pika==1.1.0
